[NTOS/CC] Fix broken usage of _SEH2_FINALLY
Finally handlers are - unlike except blocks - not part of the main function, but separate "funclets", which are called asynchronously during unwind. PSEH implements them using nested functions on GCC. While "return" from a finally handler is allowed with native SEH, it's handled by the compiler through an extra unwinding operation using _local_unwind, WHICH IS NOT SUPPORTED BY PSEH!
With PSEH, returning from a finally handler does not return from the function, instead it will only return from the finally handler and the function will continue below the finally handler as if there was no return at all.
Also: While variable assignment is handled properly with MSVC, GCC/PSEH does not know anything about exceptions and does not guarantee that your variables are set where you think they are set, because the compiler will happily remove any assignments that it considers redundant or move them to a different place. If you need to check a variable within a finally or except block, which is set in a try block, you must use _SEH2_VOLATILE.
To work around this issue, use _SEH2_EXCEPT instead, which also makes the use of _SEH2_VOLATILE obsolete, because the Result variable is set explicitly in the except handler and only checked outside of SEH.